WIP: Update MRI upstreams for Phosphorus
Adopt latest versions, namely;
- odlparent-9.0.2
- yangtools-7.0.2-SNAPSHOT
- infrautils-2.0.2
- mdsal-8.0.0-SNAPSHOT
- controller-4.0.0-SNAPSHOT
- aaa-0.14.0-SNAPSHOT
- netconf-2.0.0-SNAPSHOT

Also introduce model-{inventory,topology}, which were removed from
upstream.

Since these versions also force the Uint8/Uint32 vs. Short/Long
equivalence issue, we mass-migrate all APIs to use Uint constructs.
This in turn requires minor adjustments to mocks, as we no longer
get automatic smart zeroes.

+++ openflowjava/openflow-protocol-impl/src/main/java/org/opendaylight/openflowjava/protocol/impl/deserialization/DeserializationFactory.java
@@ -15,6 +15,7 @@ import org.opendaylight.openflowjava.protocol.api.extensibility.OFDeserializer;
 import org.opendaylight.openflowjava.protocol.api.keys.MessageCodeKey;
 import org.opendaylight.openflowjava.protocol.api.keys.TypeToClassKey;
 import org.opendaylight.yangtools.yang.binding.DataObject;
+import org.opendaylight.yangtools.yang.common.Uint8;
 
 /**
  * Factory for deserialization.
@@ -44,7 +45,7 @@ public class DeserializationFactory {
      *            version decoded from OpenFlow protocol message
      * @return correct POJO as DataObject
      */
-    public DataObject deserialize(final ByteBuf rawMessage, final short version) {
+    public DataObject deserialize(final ByteBuf rawMessage, final Uint8 version) {
         DataObject dataObject = null;
         int type = rawMessage.readUnsignedByte();
         Class<?> clazz = messageClassMap.get(new TypeToClassKey(version, type));
