+++ openflowplugin-impl/src/main/java/org/opendaylight/openflowplugin/impl/protocol/serialization/util/InstructionUtil.java
@@ -5,7 +5,6 @@
  * terms of the Eclipse Public License v1.0 which accompanies this distribution,
  * and is available at http://www.eclipse.org/legal/epl-v10.html
  */
-
 package org.opendaylight.openflowplugin.impl.protocol.serialization.util;
 
 import io.netty.buffer.ByteBuf;
@@ -14,6 +13,7 @@ import org.opendaylight.openflowjava.protocol.api.extensibility.OFSerializer;
 import org.opendaylight.openflowjava.protocol.api.extensibility.SerializerRegistry;
 import org.opendaylight.openflowjava.protocol.api.keys.MessageTypeKey;
 import org.opendaylight.yang.gen.v1.urn.opendaylight.flow.types.rev131026.instruction.Instruction;
+import org.opendaylight.yangtools.yang.common.Uint8;
 
 /**
  * Utility class for instruction serialization.
@@ -33,17 +33,13 @@ public final class InstructionUtil {
      */
     @SuppressWarnings("unchecked")
     public static void writeInstruction(final Instruction instruction,
-                                        final short version,
+                                        final Uint8 version,
                                         final SerializerRegistry registry,
                                         final ByteBuf outBuffer) {
-
         registry.<Instruction, OFSerializer<Instruction>>getSerializer(
-            new MessageTypeKey<>(
-                version,
-                (Class<Instruction>) instruction.implementedInterface()))
+            new MessageTypeKey<>(version, (Class<Instruction>) instruction.implementedInterface()))
             .serialize(instruction, outBuffer);
     }
-
 
     /**
      * Serialize instruction header.
@@ -55,15 +51,11 @@ public final class InstructionUtil {
      */
     @SuppressWarnings("unchecked")
     public static void writeInstructionHeader(final Instruction instruction,
-                                              final short version,
+                                              final Uint8 version,
                                               final SerializerRegistry registry,
                                               final ByteBuf outBuffer) {
-
         registry.<Instruction, HeaderSerializer<Instruction>>getSerializer(
-            new MessageTypeKey<>(
-                version,
-                (Class<Instruction>) instruction.implementedInterface()))
+            new MessageTypeKey<>(version, (Class<Instruction>) instruction.implementedInterface()))
             .serializeHeader(instruction, outBuffer);
     }
-
 }
